+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Assertion in 'assertGraphqlSent' helper.
 - Loading spinner.
+- Test helper class `ServiceTestCase`.
 
 ## [0.25.0] - 2024-02-05